rename a test file
