--- src/test/kotlin/top/AppTest.kt
@@ -1,13 +1,0 @@
-/*
- * This Kotlin source file was generated by the Gradle 'init' task.
- */
-package top
-
-import kotlin.test.Test
-
-class MainTest {
-    @Test
-    fun `sample test`() {
-        println("TOP should have a greeting")
-    }
-}
